some voice things that aren't really working

+++ Infrastructure/Voice/MIN.Voice.Handlers/VoiceCallParticipantLeftHandler.cs
@@ -44,7 +44,10 @@ internal sealed class VoiceCallParticipantLeftHandler : IMessageHandler
         var subRoomId = voiceParticipantLeftMessage.SubRoomId;
         var participant = voiceParticipantLeftMessage.Participant;
 
-        voicePlaybackService.RemoveParticipant(participant.Id);
+        if (voicePlaybackService.IsInVoiceCall(subRoomId))
+        {
+            voicePlaybackService.RemoveParticipant(participant.Id);
+        }
 
         await eventBus.PublishAsync(new VoiceParticipantLeftEvent()
         {
